+++ src/api/index.ts
@@ -1,4 +1,6 @@
 export * from './createFolderVo';
+export * from './createRecordVo';
+export * from './createS3UploadVo';
 export * from './getAllArchiveVos';
 export * from './getArchiveRootFolderVo';
 export * from './getFolderVo';