+++ src/types/api/FileVo.ts
@@ -8,9 +8,10 @@ export interface FileVo extends BaseVo {
   fileId: number;
   size: number;
   format: string;
-  md5Checksum: string;
-  fileURL: string;
-  downloadURL: string;
+  contentType?: string;
+  md5Checksum?: string;
+  fileURL?: string;
+  downloadURL?: string;
 }
 
 export const fileVoSchema: JSONSchemaType<FileVo> = {
@@ -25,14 +26,21 @@ export const fileVoSchema: JSONSchemaType<FileVo> = {
     format: {
       type: 'string',
     },
+    contentType: {
+      type: 'string',
+      nullable: true,
+    },
     md5Checksum: {
       type: 'string',
+      nullable: true,
     },
     fileURL: {
       type: 'string',
+      nullable: true,
     },
     downloadURL: {
       type: 'string',
+      nullable: true,
     },
     createdDT: {
       type: 'string',
@@ -45,9 +53,6 @@ export const fileVoSchema: JSONSchemaType<FileVo> = {
     'fileId',
     'size',
     'format',
-    'md5Checksum',
-    'fileURL',
-    'downloadURL',
     'createdDT',
     'updatedDT',
   ],