+++ src/types/sdk/File.ts
@@ -2,11 +2,12 @@ import type { DerivativeType } from './DerivativeType';
 
 export interface File {
   id: number;
-  derivativeType: DerivativeType;
-  fileUrl: string;
-  downloadUrl: string;
-  checksum: string;
   size: number;
+  contentType: string;
+  readonly derivativeType: DerivativeType;
+  readonly fileUrl: string;
+  readonly downloadUrl: string;
+  readonly checksum: string;
   readonly createdAt: Date;
   readonly updatedAt: Date;
 }