+++ src/utils/folderVoToFolder.ts
@@ -1,9 +1,6 @@
-import { ValidationError } from '../errors';
 import {
   isFolderVo,
-  isFolderVoArray,
   isRecordVo,
-  isRecordVoArray,
 } from '../types';
 import type {
   FolderVo,
@@ -13,29 +10,17 @@ import type {
 import { recordVoToArchiveRecord } from './recordVoToArchiveRecord';
 import { formatTimestampAsUtc } from './formatTimestampAsUtc';
 
-const extractFolderVos = (items: unknown[]): FolderVo[] => {
-  const folderVos = items.filter((item) => isFolderVo(item));
-  if (isFolderVoArray(folderVos)) {
-    return folderVos;
-  }
+const extractFolderVos = (items: unknown[]): FolderVo[] => (
+  items.filter<FolderVo>(
+    (item): item is FolderVo => isFolderVo(item),
+  )
+);
 
-  throw new ValidationError(
-    'FolderVos were improperly parsed',
-    isFolderVoArray.errors,
-  );
-};
-
-const extractRecordVos = (items: unknown[]): RecordVo[] => {
-  const recordVos = items.filter((item) => isRecordVo(item));
-  if (isRecordVoArray(recordVos)) {
-    return recordVos;
-  }
-
-  throw new ValidationError(
-    'RecordVos were improperly parsed',
-    isRecordVoArray.errors,
-  );
-};
+const extractRecordVos = (items: unknown[]): RecordVo[] => (
+  items.filter<RecordVo>(
+    (item): item is RecordVo => isRecordVo(item),
+  )
+);
 
 export const folderVoToFolder = (folderVo: FolderVo): Folder => ({
   id: folderVo.folderId,
@@ -44,7 +29,9 @@ export const folderVoToFolder = (folderVo: FolderVo): Folder => ({
   size: 0,
   createdAt: new Date(formatTimestampAsUtc(folderVo.createdDT)),
   updatedAt: new Date(formatTimestampAsUtc(folderVo.updatedDT)),
-  displayDate: new Date(formatTimestampAsUtc(folderVo.displayDT)),
+  displayDate: folderVo.displayDT !== undefined
+    ? new Date(formatTimestampAsUtc(folderVo.displayDT))
+    : undefined,
   folders: extractFolderVos(folderVo.ChildItemVOs).map(folderVoToFolder),
   archiveRecords: extractRecordVos(folderVo.ChildItemVOs).map(recordVoToArchiveRecord),
 });