+++ src/utils/index.ts
@@ -1,4 +1,6 @@
 export * from './archiveVoToArchive';
+export * from './executePresignedPost';
+export * from './fileVoToFile';
 export * from './folderVoToFolder';
 export * from './formatTimestampAsUtc';
 export * from './generateEnumTypeguard';